interface/requirements.txt
fastapi>=0.104.0,<1
uvicorn[standard]>=0.24.0,<1
httpx>=0.28.1,<1
PyYAML>=6,<7
bcrypt>=5,<6
PyJWT>=2.11,<3
python-multipart>=0.0.20,<1
